clean code and fix general issues

+++ simulator/spaces.py
@@ -1,6 +1,7 @@
 import numpy as np
 from gymnasium import spaces
-from simulator.configs.config import EnvConfig, AgentConfig
+
+from configs.config import AgentConfig, EnvConfig
 
 
 def get_single_observation_space(
@@ -12,20 +13,20 @@ def get_single_observation_space(
                 low=0,
                 high=1,
                 shape=(1, agent_config.length_view, AgentConfig.length_view),
-                dtype=np.float32,
+                dtype=np.float64,
             ),
             "goal_relative_position": spaces.Box(
                 low=-max(env_config.width, env_config.height),
                 high=max(env_config.width, env_config.height),
                 shape=(2,),
-                dtype=np.float32,
+                dtype=np.float64,
             ),
             "motion": spaces.Box(
                 low=np.array([agent_config.v_min, agent_config.omega_min]),
                 high=np.array([agent_config.v_max, agent_config.omega_max]),
-                dtype=np.float32,
+                dtype=np.float64,
             ),
-            "orientation": spaces.Box(low=-1, high=1, shape=(2,), dtype=np.float32),
+            "orientation": spaces.Box(low=-1, high=1, shape=(2,), dtype=np.float64),
         }
     )
 
@@ -34,5 +35,5 @@ def get_single_action_space(env_config: EnvConfig) -> spaces.Box:
     return spaces.Box(
         low=np.array([env_config.v_min_allowed, env_config.omega_min_allowed]),
         high=np.array([env_config.v_max_allowed, env_config.omega_max_allowed]),
-        dtype=np.float32,
+        dtype=np.float64,
     )
